refactor: lift count state and counter function to App.tsx
